blockchain: Move unique coinbase func to validate.
This moves the checkCoinbaseUniqueHeight function from accept.go to
validate.go where it more naturally belongs as it is a validation
function.  This is also evidenced by only being called from validation
functions as well.
